requirements.txt
Flask==3.1.2
Flask-SQLAlchemy==3.1.1
Flask-Bcrypt==1.0.1
mysql-connector-python==9.4.0